scripts/requirements.txt
pandas
faker
google-genai
google-cloud-bigquery
pydantic
